feat: add license lines to interface contracts

+++ contracts/interfaces/IUniswapV2Factory.sol
@@ -1,23 +1,35 @@
+// SPDX-License-Identifier: MIT
+
 pragma solidity >=0.5.0;
 
 // https://github.com/Uniswap/uniswap-v2-core/blob/master/contracts/interfaces/IUniswapV2Factory.sol
 
 interface IUniswapV2Factory {
-  event PairCreated(address indexed token0, address indexed token1, address pair, uint);
+    event PairCreated(
+        address indexed token0,
+        address indexed token1,
+        address pair,
+        uint256
+    );
 
-  function feeTo() external view returns (address);
+    function feeTo() external view returns (address);
 
-  function feeToSetter() external view returns (address);
+    function feeToSetter() external view returns (address);
 
-  function getPair(address tokenA, address tokenB) external view returns (address pair);
+    function getPair(address tokenA, address tokenB)
+        external
+        view
+        returns (address pair);
 
-  function allPairs(uint) external view returns (address pair);
+    function allPairs(uint256) external view returns (address pair);
 
-  function allPairsLength() external view returns (uint);
+    function allPairsLength() external view returns (uint256);
 
-  function createPair(address tokenA, address tokenB) external returns (address pair);
+    function createPair(address tokenA, address tokenB)
+        external
+        returns (address pair);
 
-  function setFeeTo(address) external;
+    function setFeeTo(address) external;
 
-  function setFeeToSetter(address) external;
+    function setFeeToSetter(address) external;
 }
